# Compression optimizations