--- a/lib/binding.js
+++ b/lib/binding.js
@@ -83,6 +83,9 @@ var
 
 ,	hasher 			= uint32(2654435761)
 
+,	hashTable		= new Int16Array(hashSize)
+,	emptyTable		= new Int16Array(hashSize)
+
 // CompressBound returns the maximum length of a lz4 block, given it's uncompressed length
 exports.compressBound = function (isize) {
 	return isize > maxInputSize
@@ -93,17 +96,13 @@ exports.compressBound = function (isize) {
 exports.compressHC = exports.compress
 
 exports.compress = function (src, dst, sIdx, eIdx) {
-	// V8 optimization: non sparse array with integers
-	var hashTable = new Array(hashSize)
-	for (var i = 0; i < hashSize; i++) {
-		hashTable[i] = 0
-	}
-	return compressBlock(src, dst, 0, hashTable, sIdx || 0, eIdx || dst.length)
+	hashTable.set(emptyTable)
+	return compressBlock(src, dst, 0, sIdx || 0, eIdx || dst.length)
 }
 
 exports.compressDependent = compressBlock
 
-function compressBlock (src, dst, pos, hashTable, sIdx, eIdx) {
+function compressBlock (src, dst, pos, sIdx, eIdx) {
 	var Hash = uint32() // Reusable unsigned 32 bits integer
 	var dpos = sIdx
 	var dlen = eIdx - sIdx