First commit uses typed arrays for the hashTable, second optimizes the hash computation itself.