@kripken 2015-09-12
Not sure why you're seeing something different than me, except that I'm on `node-v0.10.25`. I get `40.18` before my patches, `102` with the first, and `110` with the second.